% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnc_functions.R
\name{is.sequence.default}
\alias{is.sequence.default}
\title{Setting the default value for is.sequence methods.}
\usage{
\method{is.sequence}{default}(x, ...)
}
\arguments{
\item{x}{A vector of integers or numbers}
}
\value{
A logical value
}
\description{
Setting the default value for is.sequence methods.
}
