% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnc_functions.R
\name{is.sequence.integer}
\alias{is.sequence.integer}
\title{A function to test the "evenness" of a sequence.}
\usage{
\method{is.sequence}{integer}(x, ...)
}
\arguments{
\item{x}{A vector of integers}
}
\value{
A logical value
}
\description{
A function to test the "evenness" of a sequence.
}
