% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnc_functions.R
\name{is.sequence.numeric}
\alias{is.sequence.numeric}
\title{A function to test the "evenness" of a sequence.}
\usage{
\method{is.sequence}{numeric}(x, tol = sqrt(.Machine$double.eps), ...)
}
\arguments{
\item{x}{A vector of numbers}
}
\value{
A logical value
}
\description{
A function to test the "evenness" of a sequence.
}
