% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnc_functions.R
\name{is.sequence}
\alias{is.sequence}
\title{A function to test the "evenness" of a sequence.}
\usage{
is.sequence(x, ...)
}
\arguments{
\item{x}{A vector of integers or number}
}
\value{
A logical value
}
\description{
A function to test the "evenness" of a sequence.
}
