% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnc_functions.R
\name{is_sequence}
\alias{is_sequence}
\title{Test the "evenness" of a sequence.}
\usage{
is_sequence(x, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{A vector of integers or numbers.}

\item{tol}{Tolerance to use when checking that sequence steps are identical}
}
\value{
Logical (whether the sequence is even)
}
\description{
A sequence is "even" if it contains no \code{NA} or infinite values, and the
interval between each successive element is (approximately) identical.
}
\examples{
# Even sequence
is_sequence(c(1, 2, 3, 4, 5))

# Uneven sequence
is_sequence(c(1, 3, 5, 7, 9))

# Sequence with gaps
is_sequence(c(1, 2, 4, 5, 6))
}
