% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnc_functions.R
\name{vnc_scree}
\alias{vnc_scree}
\title{Produce a scree plot based on the VNC algorithm}
\usage{
vnc_scree(time, values, distance.measure = c("sd", "cv"))
}
\arguments{
\item{time}{A vector of sequential time intervals like years or decades}

\item{values}{A vector containing normaized frequency counts}

\item{distance.measure}{Indicating whether the standard deviation or
coefficient of variation should be used in distance calculations}
}
\value{
A scree plot
}
\description{
Produce a scree plot based on the VNC algorithm
}
