% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{witch_hunt}
\alias{witch_hunt}
\title{Counts of the phrase "witch hunt" aggregated from Google Books}
\format{
A data frame containing 17 observations of 4 variables:
\describe{
\item{decade}{Decade}
\item{token_count}{Number of uses of "witch hunt" and variants}
\item{total_count}{Total number of tokens in Google Books corpus in
this decade}
\item{counts_permil}{Uses of "witch hunt" per million tokens}
}
}
\source{
Google Books Ngram Viewer Datasets, at
\url{https://storage.googleapis.com/books/ngrams/books/datasetsv3.html}
}
\usage{
data(witch_hunt)
}
\description{
Gives the usage of "witch hunt" in each decade for which there is Google
Books data. The data is lemmatized (including both singular and plural)
and case insensitive (including, for example, "witch hunt" and "WITCH
HUNTS").
}
\examples{
data(witch_hunt)
head(witch_hunt)
}
\keyword{datasets}
