% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnc_functions.R
\name{is.sequence}
\alias{is.sequence}
\title{Test the "evenness" of a sequence.}
\usage{
is.sequence(x, tol = sqrt(.Machine$double.eps), ...)
}
\arguments{
\item{x}{A vector of integers or numbers.}

\item{tol}{Tolerance to use when checking that sequence steps are identical}

\item{...}{additional parameters passed to default sequence method.}
}
\value{
Logical (whether the sequence is even)
}
\description{
A sequence is "even" if it contains no \code{NA} or infinite values, and the
interval between each successive element is (approximately) identical.
}
