% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnc_functions.R
\name{vnc_scree}
\alias{vnc_scree}
\title{Produce a scree plot based on the VNC algorithm}
\usage{
vnc_scree(time, values, distance_measure = c("sd", "cv"))
}
\arguments{
\item{time}{A vector of sequential time intervals like years or decades}

\item{values}{A vector containing normalized frequency counts}

\item{distance_measure}{Indicating whether the standard deviation or
coefficient of variation should be used in distance calculations}
}
\value{
A scree plot
}
\description{
Produce a scree plot based on the VNC algorithm
}
\examples{
\dontrun{
# First filter to complete decades only (evenly spaced sequence)
wh_complete <- witch_hunt[witch_hunt$decade \%in\% seq(1800, 2000, 10), ]

# Create scree plot
vnc_scree(wh_complete$decade, wh_complete$counts_permil)
}
}
